remove robot folders
