+++ app/src/redux/pipettes/types.ts
@@ -1,8 +1,8 @@
 import type {
   LabwareDefinition2,
+  PipetteChannels,
   PipetteModelSpecs,
 } from '@opentrons/shared-data'
-import type { Pipette as ProtocolPipette } from '../robot/types'
 import type { RobotApiRequestMeta } from '../robot-api/types'
 import type {
   PipetteOffsetCalibration,
@@ -12,6 +12,32 @@ import type {
 // common types
 
 export type Mount = 'left' | 'right'
+
+export interface StatePipette {
+  // resource ID
+  _id: number
+  // robot mount instrument is installed on
+  mount: Mount
+  // number of liquid channels
+  channels: PipetteChannels
+  // user-given name of the instrument
+  // TODO: Ian + Mike 2018-11-06 This `name` is not what we now call the `name`,
+  // rather it is the full versioned `model` of the pipette, as placed in the
+  // Python Pipette's `name` field by the pipette factory functions.
+  // TLDR: this `name` needs to be renamed in a future PR to `model`
+  name: string
+  // tipracks the pipette uses during the protocol
+  // array of RPC object IDs corresponding to `_id` field in StateLabware
+  tipRacks: number[]
+  // string specified in protocol to load pipette
+  requestedAs?: string | null | undefined
+}
+
+export interface ProtocolPipette extends StatePipette {
+  probed: boolean
+  tipOn: boolean
+  modelSpecs: PipetteModelSpecs | null
+}
 
 export interface AttachedPipette {
   id: string
