--- app/src/redux/robot/index.ts
@@ -1,5 +1,0 @@
-// robot state module
-import * as constants from './constants'
-
-export * from './types'
-export { constants }
